- This compile warning was bugging the hell out of me on scripts.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@720 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Player.h
@@ -908,7 +908,7 @@ public:
     void                SetQuestLogSlot(QuestLogEntry *entry, uint32 slot);
     
     HEARTHSTONE_INLINE void         PushToRemovedQuests(uint32 questid)	{ m_removequests.insert(questid);}
-	HEARTHSTONE_INLINE uint32		GetFinishedDailiesCount() { return m_finishedDailyQuests.size(); }
+	HEARTHSTONE_INLINE uint32		GetFinishedDailiesCount() { return (uint32)m_finishedDailyQuests.size(); }
     void                AddToFinishedQuests(uint32 quest_id);
 	void				AddToFinishedDailyQuests(uint32 quest_id);
     void                EventTimedQuestExpire(Quest *qst, QuestLogEntry *qle, uint32 log_slot);
